requirements.txt
azure-identity 
azure-search-documents